restructure to make ticket info reusable on multiple components, presented active ticket information upon clicking
